# swagger.json
{
	"components": {
...
		"schemas": {
...
			"EMessageType": {
				"enum": [
					"text",
					"image",
					"file",
					"voice",
					"system",
					"poll",
					"sticker"
				],
				"type": "string"
			},
			"EMessageStatus": {
				"enum": [
					"sent",
					"delivered",
					"read"
				],
				"type": "string"
			},
			"Record_string.unknown_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"MessageDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"chatId": {
						"type": "string"
					},
					"senderId": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/EMessageType"
					},
					"status": {
						"$ref": "#/components/schemas/EMessageStatus"
					},
					"content": {
						"type": "string",
						"nullable": true
					},
					"replyToId": {
						"type": "string",
						"nullable": true
					},
					"forwardedFromId": {
						"type": "string",
						"nullable": true
					},
					"isEdited": {
						"type": "boolean"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"isPinned": {
						"type": "boolean"
					},
					"pinnedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"pinnedById": {
						"type": "string",
						"nullable": true
					},
					"stickerId": {
						"type": "string",
						"nullable": true
					},
					"encryptedContent": {
						"type": "string",
						"nullable": true
					},
					"encryptionMetadata": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Record_string.unknown_"
							}
						],
						"nullable": true
					},
					"keyboard": {
						"nullable": true
					},
					"linkPreviews": {
						"items": {
							"properties": {
								"siteName": {
									"type": "string",
									"nullable": true
								},
								"imageUrl": {
									"type": "string",
									"nullable": true
								},
								"description": {
									"type": "string",
									"nullable": true
								},
								"title": {
									"type": "string",
									"nullable": true
								},
								"url": {
									"type": "string"
								}
							},
							"required": [
								"siteName",
								"imageUrl",
								"description",
								"title",
								"url"
							],
							"type": "object"
						},
						"type": "array",
						"nullable": true
					},
					"scheduledAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"isScheduled": {
						"type": "boolean"
					},
					"selfDestructSeconds": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"selfDestructAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"sender": {
						"properties": {
							"avatarUrl": {
								"type": "string"
							},
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					},
					"replyTo": {
						"allOf": [
							{
								"$ref": "#/components/schemas/MessageDto"
							}
						],
						"nullable": true
					},
					"attachments": {
						"items": {
							"$ref": "#/components/schemas/MessageAttachmentDto"
						},
						"type": "array"
					},
					"reactions": {
						"items": {
							"properties": {
								"userIds": {
									"items": {
										"type": "string"
									},
									"type": "array"
								},
								"count": {
									"type": "number",
									"format": "double"
								},
								"emoji": {
									"type": "string"
								}
							},
							"required": [
								"userIds",
								"count",
								"emoji"
							],
							"type": "object"
						},
						"type": "array"
					},
					"mentions": {
						"items": {
							"properties": {
								"isAll": {
									"type": "boolean"
								},
								"userId": {
									"type": "string",
									"nullable": true
								}
							},
							"required": [
								"isAll",
								"userId"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"chatId",
					"senderId",
					"type",
					"status",
					"content",
					"replyToId",
					"forwardedFromId",
					"isEdited",
					"isDeleted",
					"isPinned",
					"pinnedAt",
					"pinnedById",
					"stickerId",
					"encryptedContent",
					"encryptionMetadata",
					"keyboard",
					"linkPreviews",
					"scheduledAt",
					"isScheduled",
					"selfDestructSeconds",
					"selfDestructAt",
					"createdAt",
					"updatedAt",
					"attachments",
					"reactions",
					"mentions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageAttachmentDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"fileId": {
						"type": "string"
					},
					"fileName": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"fileType": {
						"type": "string"
					},
					"fileSize": {
						"type": "number",
						"format": "double"
					},
					"thumbnailUrl": {
						"type": "string",
						"nullable": true
					},
					"width": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"height": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"id",
					"fileId",
					"fileName",
					"fileUrl",
					"fileType",
					"fileSize",
					"thumbnailUrl",
					"width",
					"height"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IBotEditMessageBody": {
				"properties": {
					"content": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IMessageSearchDto": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/MessageDto"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalCount"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IEditMessageBody": {
				"properties": {
					"content": {
						"type": "string"
					}
				},
				"required": [
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAddReactionBody": {
				"properties": {
					"emoji": {
						"type": "string"
					}
				},
				"required": [
					"emoji"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PollOptionDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"position": {
						"type": "number",
						"format": "double"
					},
					"voterCount": {
						"type": "number",
						"format": "double"
					},
					"voterIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"text",
					"position",
					"voterCount",
					"voterIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PollDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"messageId": {
						"type": "string"
					},
					"question": {
						"type": "string"
					},
					"isAnonymous": {
						"type": "boolean"
					},
					"isMultipleChoice": {
						"type": "boolean"
					},
					"isClosed": {
						"type": "boolean"
					},
					"closedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"options": {
						"items": {
							"$ref": "#/components/schemas/PollOptionDto"
						},
						"type": "array"
					},
					"totalVotes": {
						"type": "number",
						"format": "double"
					},
					"userVotedOptionIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"messageId",
					"question",
					"isAnonymous",
					"isMultipleChoice",
					"isClosed",
					"closedAt",
					"options",
					"totalVotes",
					"userVotedOptionIds",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IVotePollBody": {
				"properties": {
					"optionIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"optionIds"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "rest-api",
		"version": "1.0.1",
		"description": "REST API Template App",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/bot-api/message/{id}/edit": {
			"post": {
				"operationId": "EditMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageDto"
								}
							}
						}
					}
				},
				"description": "Редактировать сообщение бота.",
				"summary": "Bot: редактирование сообщения",
				"tags": [
					"Bot API"
				],
				"security": [
					{
						"bot": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBotEditMessageBody"
							}
						}
					}
				}
			}
		},
		"/api/bot-api/message/{id}": {
			"delete": {
				"operationId": "DeleteMessage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Удалить сообщение бота.",
				"summary": "Bot: удаление сообщения",
				"tags": [
					"Bot API"
				],
				"security": [
					{
						"bot": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/chat/{chatId}/message/search": {
			"get": {
				"operationId": "SearchMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessageSearchDto"
								}
							}
						}
					}
				},
				"description": "Поиск сообщений в чате.",
				"summary": "Поиск в чате",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "chatId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/api/message/search": {
			"get": {
				"operationId": "SearchMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessageSearchDto"
								}
							}
						}
					}
				},
				"description": "Глобальный поиск по сообщениям во всех чатах пользователя.",
				"summary": "Глобальный поиск сообщений",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/message/{id}": {
			"patch": {
				"operationId": "EditMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageDto"
								}
							}
						}
					}
				},
				"description": "Отредактировать сообщение.",
				"summary": "Редактирование сообщения",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEditMessageBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMessage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Удалить сообщение (soft delete).",
				"summary": "Удаление сообщения",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/message/{id}/reaction": {
			"post": {
				"operationId": "AddReaction",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Добавить реакцию на сообщение.",
				"summary": "Добавление реакции",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAddReactionBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveReaction",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Удалить реакцию с сообщения.",
				"summary": "Удаление реакции",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/message/{id}/pin": {
			"post": {
				"operationId": "PinMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageDto"
								}
							}
						}
					}
				},
				"description": "Закрепить сообщение.",
				"summary": "Закрепление сообщения",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "UnpinMessage",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Открепить сообщение.",
				"summary": "Открепление сообщения",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/message/{id}/open": {
			"post": {
				"operationId": "MarkMessageOpened",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MessageDto"
								}
							}
						}
					}
				},
				"description": "Отметить сообщение как открытое (запускает таймер самоуничтожения).",
				"summary": "Открытие сообщения",
				"tags": [
					"Message"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/poll/{id}/vote": {
			"post": {
				"operationId": "Vote",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PollDto"
								}
							}
						}
					}
				},
				"description": "Проголосовать в опросе.",
				"summary": "Голосование",
				"tags": [
					"Poll"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVotePollBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RetractVote",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PollDto"
								}
							}
						}
					}
				},
				"description": "Отозвать голос.",
				"summary": "Отзыв голоса",
				"tags": [
					"Poll"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/poll/{id}/close": {
			"post": {
				"operationId": "ClosePoll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PollDto"
								}
							}
						}
					}
				},
				"description": "Закрыть опрос.",
				"summary": "Закрытие опроса",
				"tags": [
					"Poll"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/poll/{id}": {
			"get": {
				"operationId": "GetPoll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PollDto"
								}
							}
						}
					}
				},
				"description": "Получить опрос по ID.",
				"summary": "Получение опроса",
				"tags": [
					"Poll"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		},
		{
			"url": "http://0.0.0.0:8181",
			"description": "Local dev"
		},
		{
			"url": "http://147.45.245.104:8181",
			"description": "Host"
		}
	]
}